acge_backend/requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
Markdown==3.4.1
mypy-extensions==0.4.3
numpy==1.23.5
pandas==1.5.2
pathspec==0.10.2
platformdirs==2.5.4
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
